backend/requirements.txt
fastapi>=0.136.0
uvicorn>=0.46.0
sqlalchemy>=2.0.0
asyncmy>=0.2.0
alembic>=1.18.0
pydantic-settings>=2.14.0
python-jose>=3.5.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0,<5.0
python-multipart>=0.0.26
httpx>=0.28.0
